Refactor server compiler sql
